update runtime on the data
the accuracy is 99.38% with the test error of 0.62%. This is not achieving the state of the art that score 0.21% (highest). Maybe we need to use different optimization than Adam and re-adjust the model. This had been trained on only 5 epochs (it can be more). The time per second to run is 211s which is slow ( because I am only using CPU).